refactor: fold RealDatabaseTest into WikiLinkFixerTests, delete stale file
RealDatabaseTest no longer tested a real database — it was inline markdown
regression cases for WikiLinkFixer.applyFixes(). Moved those cases into
WikiLinkFixerTests as testApplyFixesProductionPatterns and
testApplyFixesUnchangedWhenClean; deleted RealDatabaseTest.swift.
